+++ src/components/layout/Head.jsx
@@ -64,6 +64,17 @@ function Head({ article, author, description, errorPage, home, image, locale, pa
             }
           }
         }
+        allContentfulCurriculumVitaePage(
+          filter: { contentful_id: { eq: "29kQlzt1s2bR8OirrtTbCo" } }
+        ) {
+          edges {
+            node {
+              contentful_id
+              node_locale
+              slug
+            }
+          }
+        }
       }
     `,
   );
@@ -121,7 +132,10 @@ function Head({ article, author, description, errorPage, home, image, locale, pa
           return <meta property="og:url" content={`${baseURL}/${pagePath}`} />;
         } else {
           return (
-            <meta property="og:url" content={createLocaleURL(baseURL, pageID, locale, data)} />
+            <meta
+              property="og:url"
+              content={createLocaleURL(baseURL, pageID, locale, data).replace('//', '/')}
+            />
           );
         }
       })()}
@@ -164,7 +178,7 @@ function Head({ article, author, description, errorPage, home, image, locale, pa
           return siteMetadata.locales.map((listLocale) => (
             <link
               rel="alternate"
-              href={createLocaleURL(baseURL, pageID, listLocale, data)}
+              href={createLocaleURL(baseURL, pageID, listLocale, data).replace('//', '/')}
               hrefLang={siteMetadata.simpleLocales[listLocale.replace('-', '_')]}
               key={listLocale}
             />