+++ src/locales/en.js
@@ -3,15 +3,19 @@
 
 import { defineMessages } from 'react-intl';
 
+import cvMessages from './en/cvMessages';
 import footerMessages from './en/footerMessages';
 import headerMessages from './en/headerMessages';
 import headMessages from './en/headMessages';
+import hiddenSectionMessages from './en/hiddenSectionMessages';
 import notFoundMessages from './en/notFoundMessages';
 
 export const messages = {
+  ...cvMessages,
   ...footerMessages,
   ...headerMessages,
   ...headMessages,
+  ...hiddenSectionMessages,
   ...notFoundMessages,
 };