+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.11.0] - 2021-04-15
+
+### Added
+
+- Template component for curriculum vitae pages.
+- Curriculum vitae pages.
+- Main menu as the primary means of navigation to the header.
+- Sections that can be hidden and shown with a button.
+
+### Changed
+
+- Style of the button components to have an underline instead of the background colour.
+
+### Removed
+
+- Unused colours from the theme.
+
 ## [0.10.1] - 2021-04-12
 
 ### Fixed
@@ -345,7 +362,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - Existing contents of the site from the old source code repository at [anttikivi/shared-sonnet](https://github.com/anttikivi/shared-sonnet).
 
-[unreleased]: https://github.com/anttikivi/anttikivi.fi/compare/v0.10.1...HEAD
+[unreleased]: https://github.com/anttikivi/anttikivi.fi/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/anttikivi/anttikivi.fi/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/anttikivi/anttikivi.fi/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/anttikivi/anttikivi.fi/compare/v0.9.2...v0.10.0
 [0.9.2]: https://github.com/anttikivi/anttikivi.fi/compare/v0.9.1...v0.9.2